Enhancement

+++ contracts/stub/MockRegistry.sol
@@ -6,9 +6,13 @@ import "@chainlink/contracts/src/v0.8/interfaces/FeedRegistryInterface.sol";
 
 contract MockRegistry {
     uint8 private _decimals = 8;
-    mapping(address => mapping(address => int)) private _price;
+    mapping(address => mapping(address => int256)) private _price;
 
-    function decimals(address base, address quote) external view returns (uint8) {
+    function decimals(address base, address quote)
+        external
+        view
+        returns (uint8)
+    {
         // Shh
         base;
         quote;
@@ -16,9 +20,13 @@ contract MockRegistry {
         return _decimals;
     }
 
-     function getFeed(address base, address quote) external view returns (AggregatorV2V3Interface) {
+    function getFeed(address base, address quote)
+        external
+        view
+        returns (AggregatorV2V3Interface)
+    {
         // Shh
-         _price;
+        _price;
         base;
         quote;
 
@@ -33,7 +41,17 @@ contract MockRegistry {
         return true;
     }
 
-    function latestRoundData(address base, address quote) external view returns (uint80, int, uint, uint, uint80) {
+    function latestRoundData(address base, address quote)
+        external
+        view
+        returns (
+            uint80,
+            int256,
+            uint256,
+            uint256,
+            uint80
+        )
+    {
         // Shh
         base;
         quote;
@@ -41,7 +59,11 @@ contract MockRegistry {
         return (0, _price[base][quote], 0, 0, 0);
     }
 
-    function setPrice(address base, address quote, int price) external {
+    function setPrice(
+        address base,
+        address quote,
+        int256 price
+    ) external {
         _price[base][quote] = price;
     }
